+++ lib/src/commands/check/format.dart
@@ -10,6 +10,7 @@ import 'package:gg_one/src/tools/formatter.dart';
 import 'package:gg_lang/gg_lang.dart';
 import 'package:gg_args/gg_args.dart';
 import 'package:gg_log/gg_log.dart';
+import 'package:gg_status_printer/gg_status_printer.dart';
 import 'package:mocktail/mocktail.dart' as mocktail;
 
 // #############################################################################
@@ -41,9 +42,17 @@ class Format extends DirCommand<void> {
 
     final type = checkProjectType(directory);
 
+    if (type == ProjectType.none) {
+      GgStatusPrinter<void>(
+        ggLog: ggLog,
+        message: 'Skipping format (no project manifest)',
+      ).logStatus(GgStatusPrinterStatus.success);
+      return;
+    }
+
     final formatter = switch (type) {
       ProjectType.dart || ProjectType.flutter => _dartFormatter,
-      ProjectType.typescript => _typeScriptFormatter,
+      ProjectType.typescript || ProjectType.none => _typeScriptFormatter,
     };
 
     await formatter.run(directory: directory, ggLog: ggLog);